Created dummy structure for consumer
