+++ pom.xml
@@ -4,7 +4,7 @@
 	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 	<modelVersion>4.0.0</modelVersion>
 
-	<groupId>di-kafkameter</groupId>
+	<groupId>com.di.jmeter.kafka</groupId>
 	<artifactId>di-kafkameter</artifactId>
 	<version>1.1</version>
 
@@ -13,66 +13,48 @@
 
 	<properties>
 		<jdk.version>1.8</jdk.version>
-		<di-jmeter-version>5.2.1</di-jmeter-version>
+		<jmeter-version>5.5</jmeter-version>
+		<kafka.client.version>3.3.1</kafka.client.version>
 		<jmeter.lib.scope>provided</jmeter.lib.scope>
-		<kafka.client.version>2.5.0</kafka.client.version>
+
 	</properties>
 
 	<dependencies>
 		<dependency>
 			<groupId>org.apache.jmeter</groupId>
 			<artifactId>ApacheJMeter_java</artifactId>
-			<version>${di-jmeter-version}</version>
+			<version>${jmeter-version}</version>
 			<scope>${jmeter.lib.scope}</scope>
-			<exclusions>
-				<exclusion>
-					<groupId>com.google.guava</groupId>
-					<artifactId>guava</artifactId>
-				</exclusion>
-			</exclusions>
 		</dependency>
-
-		<dependency>
-			<groupId>com.google.guava</groupId>
-			<artifactId>guava</artifactId>
-			<version>14.0</version>
-		</dependency>
-
-		<dependency>
-			<groupId>com.google.code.findbugs</groupId>
-			<artifactId>jsr305</artifactId>
-			<version>1.3.9</version>
-		</dependency>
-
-		<dependency>
-			<groupId>com.google.code.gson</groupId>
-			<artifactId>gson</artifactId>
-			<version>2.2.4</version>
-		</dependency>
-
-		<!-- Kafka Dependencies -->
-
 		<dependency>
 			<groupId>org.apache.kafka</groupId>
 			<artifactId>kafka-clients</artifactId>
 			<version>${kafka.client.version}</version>
-			<exclusions>
-				<!-- XXX Transitively includes log4j 1.2.15 which has bad metadata http://stackoverflow.com/a/9047963 -->
-				<exclusion>
-					<groupId>com.sun.jmx</groupId>
-					<artifactId>jmxri</artifactId>
-				</exclusion>
-				<exclusion>
-					<groupId>com.sun.jdmk</groupId>
-					<artifactId>jmxtools</artifactId>
-				</exclusion>
-				<exclusion>
-					<groupId>javax.jms</groupId>
-					<artifactId>jms</artifactId>
-				</exclusion>
-			</exclusions>
+		</dependency>
+		<dependency>
+			<groupId>org.slf4j</groupId>
+			<artifactId>slf4j-api</artifactId>
+			<version>1.7.29</version>
+		</dependency>
+<!--		<dependency>-->
+<!--			<groupId>com.google.code.gson</groupId>-->
+<!--			<artifactId>gson</artifactId>-->
+<!--			<version>2.2.4</version>-->
+<!--		</dependency>-->
+		<!-- https://mvnrepository.com/artifact/com.google.guava/guava -->
+		<dependency>
+			<groupId>com.google.guava</groupId>
+			<artifactId>guava</artifactId>
+			<version>31.1-jre</version>
 		</dependency>
 
+<!--		<dependency>-->
+<!--			<groupId>com.google.code.findbugs</groupId>-->
+<!--			<artifactId>jsr305</artifactId>-->
+<!--			<version>1.3.9</version>-->
+<!--		</dependency>-->
+
+		<!-- Kafka Dependencies -->
 	</dependencies>
 
 	<build>
