+++ src/main/java/com/di/jmeter/kafka/config/KafkaProducerConfigResources.properties
@@ -13,7 +13,5 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-#
-# DUMMY Property file to avoid from getting errors
 
 displayName=Kafka Config
