+++ src/main/java/com/di/jmeter/kafka/sampler/KafkaConsumerSamplerResources.properties
@@ -14,4 +14,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-displayName=Kafka producer sampler
+displayName=Kafka consumer sampler
